third_party/blink/web_tests/wpt_internal/app-history/navigate-event/respondWith-multiple-times.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  appHistory.onnavigate = t.step_func_done(e => {
    e.respondWith(Promise.resolve());
    assert_throws_dom("InvalidStateError", () => e.respondWith(Promise.resolve()));
  });

  location.href = "#1";
  assert_equals(location.hash, "");
}, "AppHistoryNavigateEvent.respondWith throws if invoked more than once");
</script>
